update controller methods

+++ backend/src/main/java/saho/backend/service/EmployeeService.java
@@ -1,10 +1,12 @@
 package saho.backend.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import saho.backend.exeption.ResourceNotFoundException;
 import saho.backend.model.Employee;
 import saho.backend.repo.EmployeeRepo;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -23,13 +25,33 @@ public class EmployeeService {
                 .orElseThrow(() -> new RuntimeException("Not Found"));
     }
 
+
     public Employee saveEmployee(Employee employee) {
         return employeeRepo.save(employee);
     }
 
     public void deleteEmployee(String id) {
-        employeeRepo.deleteById(id);
+        Optional<Employee> employeeOptional = employeeRepo.findById(id);
+        if (employeeOptional.isPresent()) {
+            employeeRepo.deleteById(id);
+        } else {
+            throw new ResourceNotFoundException("Employee not found with id: " + id);
+        }
     }
+
+    public Employee updateEmployee(String id, Employee employeeDetails) {
+        Employee existingEmployee = employeeRepo.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Employee not found with id: " + id));
+
+        existingEmployee.setFirstName(employeeDetails.getFirstName());
+        existingEmployee.setLastName(employeeDetails.getLastName());
+        existingEmployee.setEmailId(employeeDetails.getEmailId());
+
+        return employeeRepo.save(existingEmployee);
+    }
+
+
+
